// CapstoneProjectFrontend/src/app/enums/audit-action.enum.ts
export enum AuditAction {
    UserRegistered = 'UserRegistered',
    UserLoggedIn = 'UserLoggedIn',
    DocumentUploaded = 'DocumentUploaded',
    NewVersionUploaded = 'NewVersionUploaded',
    DocumentApproved = 'DocumentApproved',
    DocumentRejected = 'DocumentRejected',
    DocumentDownloaded = 'DocumentDownloaded',
    DocumentForwarded = 'DocumentForwarded',
    DocumentWithdrawn = 'DocumentWithdrawn',
    UserDeactivated = 'UserDeactivated',
    UserReactivated = 'UserReactivated',
    DepartmentDeleted = 'DepartmentDeleted',
    PasswordChanged = 'PasswordChanged'
}